Dynamic: include updateProps in doc API?
It seems like `props.updateProps` is used inside the `Dynamic` component but is not mentioned in the documentation: https://idyll-lang.org/docs/components/default/dynamic

I'm not sure if this is intentional though. Maybe you don't want people to pass in callbacks when writing markdown? 

My motivation for adding it was that I want to reuse the Dynamic React component directly (not in markdown)