Template projects: Add link markup around bare URL to prevent 404 link
The template pages in this commit each contain a bare URL pointing to
the idyll docs. The idyll compiler compiles these into links. But since
there is a comma adjacent to the end of each link, the idyll compiler
includes the comma in the linked-to URL, causing a 404 upon clicking the
link.

This commit replaces each `bare url,` with `[bare url](bare url),` to
prevent this behavior.

Github issue: https://github.com/idyll-lang/idyll/issues/614